Larvel Authorization

+++ app/Http/Controllers/CommentController.php
@@ -4,14 +4,22 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function add()
     {
         $comment = new Comment();
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -19,8 +27,15 @@ class CommentController extends Controller
 
     public function delete($id)
     {
+
         $comment = Comment::find($id);
-        $comment->delete();
+
+        if(Gate::allows('comment-delete', $comment)){
+            $comment->delete();
+            return back()->with('DelComSuccess', 'Delete Comment Successfully');
+        }else {
+            return back()->with("error", "Unanthorize");
+        }
 
         return back();
     }
